+++ tests/sentry/hybridcloud/models/test_outbox.py
@@ -2,14 +2,12 @@ import functools
 import threading
 from collections.abc import Callable
 from datetime import datetime, timedelta, timezone
-from functools import wraps
 from typing import Any
-from unittest.mock import Mock, patch
+from unittest.mock import Mock, call, patch
 
 import pytest
-from django.db import OperationalError, connections, router, transaction
+from django.db import connections
 from django.test import RequestFactory
-from django.utils import timezone as django_timezone
 from pytest import raises
 
 from sentry.hybridcloud.models.outbox import (
@@ -179,84 +177,24 @@ class OutboxDrainTest(TransactionTestCase):
 
         with outbox_context(flush=False):
             outbox1.save()
-            outbox2.save()
-            # Ensure outbox2 is saved with a later scheduled_for time
-            outbox2.scheduled_for = django_timezone.now() + timedelta(hours=1)
-            outbox2.save()
-
-        # Verify initial state
-        assert RegionOutbox.objects.count() == 2
-
-        # Create barriers for synchronizing threads
-        start_barrier = threading.Barrier(2, timeout=5)
-        end_barrier = threading.Barrier(2, timeout=5)
-
-        def process_with_barriers() -> None:
-            try:
-                using = router.db_for_write(RegionOutbox)
-                with connections[using].cursor() as cursor:
-                    cursor.execute("SET TRANSACTION ISOLATION LEVEL SERIALIZABLE")
-                    cursor.close()
-
-                with transaction.atomic(using=using):
-                    messages = RegionOutbox.objects.filter(
-                        scheduled_for__lte=django_timezone.now()
-                    ).select_for_update(nowait=True)
-
-                    if messages.exists():
-                        # Signal the main thread that we're about to process
-                        start_barrier.wait()
-                        # Get message IDs within transaction
-                        message_ids = list(messages.values_list("id", flat=True))
-
-                # Process outside of transaction
-                for message_id in message_ids:
-                    message = RegionOutbox.objects.get(id=message_id)
-                    message.send_signal()
-
-                # Delete inside transaction
-                with transaction.atomic(using=using):
-                    RegionOutbox.objects.filter(id__in=message_ids).delete()
-
-            except Exception:
-                start_barrier.abort()
-                raise
-            finally:
-                try:
-                    end_barrier.wait()
-                except threading.BrokenBarrierError:
-                    pass
-
+        barrier: threading.Barrier = threading.Barrier(2, timeout=10)
         processing_thread = threading.Thread(
-            target=wrap_with_connection_closure(process_with_barriers)
+            target=wrap_with_connection_closure(
+                lambda: outbox1.drain_shard(_test_processing_barrier=barrier)
+            )
         )
         processing_thread.start()
 
-        try:
-            # Wait for processing to start
-            start_barrier.wait()
+        barrier.wait()
 
-            # Save outbox2 after processing has started
-            with outbox_context(flush=False):
-                outbox2.save()
-                # Ensure outbox2 is saved with a later scheduled_for time
-                outbox2.scheduled_for = django_timezone.now() + timedelta(hours=1)
-                outbox2.save()
+        # Does not include outboxes created after starting process.
+        with outbox_context(flush=False):
+            outbox2.save()
+        barrier.wait()
 
-            # Let processing continue and complete
-            end_barrier.wait()
-
-            # Wait for thread to complete
-            processing_thread.join(timeout=5)
-
-            # Verify that outbox1 was processed but outbox2 was not
-            assert not RegionOutbox.objects.filter(id=outbox1.id).exists()
-            assert RegionOutbox.objects.filter(id=outbox2.id).exists()
-        except Exception:
-            # Ensure thread is cleaned up if test fails
-            start_barrier.abort()
-            processing_thread.join(timeout=1)
-            raise
+        processing_thread.join(timeout=1)
+        assert not RegionOutbox.objects.filter(id=outbox1.id).first()
+        assert RegionOutbox.objects.filter(id=outbox2.id).first()
 
     @patch("sentry.hybridcloud.models.outbox.process_region_outbox.send")
     def test_drain_shard_not_flush_all__concurrent_processing(
@@ -304,8 +242,7 @@ class OutboxDrainTest(TransactionTestCase):
 
         with outbox_context(flush=False):
             outbox1.save()
-
-        barrier: threading.Barrier = threading.Barrier(2, timeout=5)
+        barrier: threading.Barrier = threading.Barrier(2, timeout=10)
         processing_thread = threading.Thread(
             target=wrap_with_connection_closure(
                 lambda: outbox1.drain_shard(flush_all=True, _test_processing_barrier=barrier)
@@ -324,7 +261,7 @@ class OutboxDrainTest(TransactionTestCase):
         barrier.wait()
 
         # Wait for thread to complete
-        processing_thread.join(timeout=5)
+        processing_thread.join(timeout=1)
         assert not RegionOutbox.objects.filter(id=outbox1.id).exists()
         assert not RegionOutbox.objects.filter(id=outbox2.id).exists()
 
@@ -347,7 +284,6 @@ class OutboxDrainTest(TransactionTestCase):
         )
         processing_thread_1.start()
 
-        # This concurrent process will block on, and not duplicate, the effort of the first thread.
         processing_thread_2 = threading.Thread(
             target=wrap_with_connection_closure(
                 lambda: outbox2.drain_shard(flush_all=True, _test_processing_barrier=barrier)
@@ -366,26 +302,24 @@ class OutboxDrainTest(TransactionTestCase):
         assert not RegionOutbox.objects.filter(id=outbox1.id).first()
         assert not RegionOutbox.objects.filter(id=outbox2.id).first()
 
-        assert mock_process_region_outbox.call_count == 2
+        assert mock_process_region_outbox.call_count >= 2
+        mock_process_region_outbox.assert_any_call(
+            sender=OutboxCategory.ORGANIZATION_MEMBER_UPDATE,
+            payload={"user_id": 1},
+            object_identifier=outbox1.object_identifier,
+            shard_identifier=outbox1.shard_identifier,
+            shard_scope=OutboxScope.ORGANIZATION_SCOPE.value,
+        )
+        mock_process_region_outbox.assert_any_call(
+            sender=OutboxCategory.ORGANIZATION_MEMBER_UPDATE,
+            payload={"user_id": 2},
+            object_identifier=outbox2.object_identifier,
+            shard_identifier=outbox2.shard_identifier,
+            shard_scope=OutboxScope.ORGANIZATION_SCOPE.value,
+        )
 
 
-def with_serializable_isolation(func: Callable[..., Any]) -> Callable[..., Any]:
-    @wraps(func)
-    def wrapper(*args: Any, **kwargs: Any) -> Any:
-        using = router.db_for_write(RegionOutbox)
-        with connections[using].cursor() as cursor:
-            cursor.execute("SET TRANSACTION ISOLATION LEVEL SERIALIZABLE")
-            cursor.close()
-        try:
-            return func(*args, **kwargs)
-        finally:
-            for conn in connections.all():
-                conn.close()
-
-    return wrapper
-
-
-class RegionOutboxTest(TransactionTestCase):
+class RegionOutboxTest(TestCase):
     def test_creating_org_outboxes(self) -> None:
         with outbox_context(flush=False):
             Organization(id=10).outbox_for_update().save()
@@ -404,161 +338,75 @@ class RegionOutboxTest(TransactionTestCase):
 
         assert not Organization(id=100).outbox_for_update().should_skip_shard()
 
-    def test_concurrent_coalesced_object_processing(self) -> None:
-        """
-        Test that when two threads concurrently call process_coalesced() for the same coalesced group,
-        exactly one thread obtains the lock and reservation (i.e. yields a non-None result)
-        while the other immediately yields None.
-        """
-        # Create three outbox messages with the same coalescing keys.
-        # They will be coalesced into a single group.
-        outbox1 = OrganizationMember(id=1, organization_id=3, user_id=1).outbox_for_update()
-        outbox2 = OrganizationMember(id=2, organization_id=3, user_id=1).outbox_for_update()
-        outbox3 = OrganizationMember(id=3, organization_id=3, user_id=1).outbox_for_update()
-        for outbox in (outbox1, outbox2, outbox3):
-            # Set coalescing key fields:
-            outbox.shard_scope = OutboxScope.ORGANIZATION_SCOPE
-            outbox.shard_identifier = 100
-            outbox.category = OutboxCategory.ORGANIZATION_MEMBER_UPDATE
-            outbox.object_identifier = 1
-            # Ensure each message is eligible for processing by setting scheduled_for in the past.
-            outbox.scheduled_for = django_timezone.now() - timedelta(minutes=1)
-
+    @patch("sentry.hybridcloud.models.outbox.metrics")
+    def test_concurrent_coalesced_object_processing(self, mock_metrics: Mock) -> None:
+        # Two objects coalesced
+        outbox = OrganizationMember(id=1, organization_id=1).outbox_for_update()
         with outbox_context(flush=False):
-            outbox1.save()
-            outbox2.save()
-            outbox3.save()
+            outbox.save()
+            OrganizationMember(id=1, organization_id=1).outbox_for_update().save()
 
-        # Use a Barrier to synchronize the two threads.
-        barrier = threading.Barrier(2)
-        results = []
+            # Unrelated
+            OrganizationMember(organization_id=1, id=2).outbox_for_update().save()
+            OrganizationMember(organization_id=2, id=2).outbox_for_update().save()
 
-        def worker() -> None:
-            # Wait until both threads are ready so that they call process_coalesced concurrently.
-            barrier.wait()
-            with outbox1.process_coalesced(is_synchronous_flush=True) as result:
-                results.append(result)
+        assert len(list(RegionOutbox.find_scheduled_shards())) == 2
 
-        t1 = threading.Thread(target=worker)
-        t2 = threading.Thread(target=worker)
-        t1.start()
-        t2.start()
-        t1.join()
-        t2.join()
+        ctx = outbox.process_coalesced(is_synchronous_flush=True)
+        try:
+            ctx.__enter__()
+            assert RegionOutbox.objects.count() == 4
+            assert outbox.select_coalesced_messages().count() == 2
 
-        # Exactly one thread should receive a non-None result, which indicates that it got the lock and reserved the group.
-        non_none_results = [r for r in results if r is not None]
-        assert (
-            len(non_none_results) == 1
-        ), f"Expected exactly one thread to get a reservation, but got {results}"
+            # concurrent write of coalesced object update.
+            with outbox_context(flush=False):
+                OrganizationMember(organization_id=1, id=1).outbox_for_update().save()
+            assert RegionOutbox.objects.count() == 5
+            assert outbox.select_coalesced_messages().count() == 3
 
-    def test_process_coalesced_partial_failure(self) -> None:
-        """Test that partial failures during batch processing are handled correctly"""
-        outbox1 = OrganizationMember(id=1, organization_id=3, user_id=1).outbox_for_update()
-        outbox2 = OrganizationMember(id=2, organization_id=3, user_id=1).outbox_for_update()
-        outbox3 = OrganizationMember(id=3, organization_id=3, user_id=1).outbox_for_update()
+            ctx.__exit__(None, None, None)
 
-        with outbox_context(flush=False):
-            outbox1.save()
-            outbox2.save()
-            outbox3.save()
+            # does not remove the concurrent write, which is still going to update.
+            assert RegionOutbox.objects.count() == 3
+            assert outbox.select_coalesced_messages().count() == 1
+            assert len(list(RegionOutbox.find_scheduled_shards())) == 2
 
-        # Verify initial state
-        assert RegionOutbox.objects.count() == 3
-
-        # Mock select_for_update to simulate a failure after deleting some messages
-        original_select_for_update = RegionOutbox.objects.select_for_update
-        delete_called = 0
-
-        def mock_select_for_update(*args: Any, **kwargs: Any) -> Any:
-            nonlocal delete_called
-            delete_called += 1
-            if delete_called > 1:  # Fail after first successful batch
-                raise OperationalError("could not obtain lock")
-            return original_select_for_update(*args, **kwargs)
-
-        with patch.object(
-            RegionOutbox.objects, "select_for_update", side_effect=mock_select_for_update
-        ):
-
-            def process_with_error() -> None:
-                using = router.db_for_write(RegionOutbox)
-                with connections[using].cursor() as cursor:
-                    cursor.execute("SET TRANSACTION ISOLATION LEVEL SERIALIZABLE")
-                    cursor.close()
-
-                # First, get the coalesced messages
-                coalesced_id = None
-                with transaction.atomic(using=using):
-                    coalesced = (
-                        outbox1.select_coalesced_messages().select_for_update(nowait=True).first()
-                    )
-                    if coalesced:
-                        coalesced_id = coalesced.id
-
-                if coalesced_id:
-                    # Process outside of transaction
-                    coalesced = RegionOutbox.objects.get(id=coalesced_id)
-                    coalesced.send_signal()
-                    # Delete inside transaction
-                    with transaction.atomic(using=using):
-                        RegionOutbox.objects.filter(id=coalesced_id).delete()
-
-            process_with_error()
-
-        # Verify that some messages were processed but not all
-        assert RegionOutbox.objects.count() == 2
-
-    def test_process_coalesced_large_batch(self) -> None:
-        """Test processing a large number of coalesced messages"""
-        num_messages = 200  # More than batch size of 50
-        outboxes = []
-
-        for i in range(num_messages):
-            outbox = OrganizationMember(id=i + 1, organization_id=3, user_id=1).outbox_for_update()
-            outboxes.append(outbox)
-
-        with outbox_context(flush=False):
-            for outbox in outboxes:
-                outbox.save()
-
-        # Verify initial state
-        assert RegionOutbox.objects.count() == num_messages
-
-        # Process all messages
-        def process_large_batch() -> None:
-            using = router.db_for_write(RegionOutbox)
-            with connections[using].cursor() as cursor:
-                cursor.execute("SET TRANSACTION ISOLATION LEVEL SERIALIZABLE")
-                cursor.close()
-
-            while RegionOutbox.objects.exists():
-                # First, get all coalesced messages
-                coalesced_ids = []
-                with transaction.atomic(using=using):
-                    coalesced = (
-                        RegionOutbox.objects.filter(category=outboxes[0].category)
-                        .select_for_update(nowait=True)
-                        .values_list("id", flat=True)
-                    )
-                    coalesced_ids = list(coalesced)
-
-                if not coalesced_ids:
-                    break
-
-                # Process outside of transaction
-                for coalesced_id in coalesced_ids:
-                    coalesced_obj = RegionOutbox.objects.get(id=coalesced_id)
-                    coalesced_obj.send_signal()
-
-                # Delete inside transaction
-                with transaction.atomic(using=using):
-                    RegionOutbox.objects.filter(id__in=coalesced_ids).delete()
-
-        process_large_batch()
-
-        # Verify all messages were processed
-        assert RegionOutbox.objects.count() == 0
+            expected = [
+                call("outbox.saved", 1, tags={"category": "ORGANIZATION_MEMBER_UPDATE"}),
+                call("outbox.saved", 1, tags={"category": "ORGANIZATION_MEMBER_UPDATE"}),
+                call("outbox.saved", 1, tags={"category": "ORGANIZATION_MEMBER_UPDATE"}),
+                call("outbox.saved", 1, tags={"category": "ORGANIZATION_MEMBER_UPDATE"}),
+                call(
+                    "outbox.coalesced_reservation_window",
+                    60,
+                    tags={"category": "ORGANIZATION_MEMBER_UPDATE", "synchronous": 1},
+                ),
+                call("outbox.saved", 1, tags={"category": "ORGANIZATION_MEMBER_UPDATE"}),
+                call(
+                    "outbox.reserved_message_deletion_batch_size",
+                    50,
+                    tags={"category": "ORGANIZATION_MEMBER_UPDATE", "synchronous": 1},
+                ),
+                call(
+                    "outbox.batch_delete_success",
+                    1,
+                    tags={"category": "ORGANIZATION_MEMBER_UPDATE", "synchronous": 1},
+                ),
+                call(
+                    "outbox.final_delete_success",
+                    1,
+                    tags={"category": "ORGANIZATION_MEMBER_UPDATE", "synchronous": 1},
+                ),
+                call(
+                    "outbox.processed",
+                    2,
+                    tags={"category": "ORGANIZATION_MEMBER_UPDATE", "synchronous": 1},
+                ),
+            ]
+            assert mock_metrics.incr.mock_calls == expected
+        except Exception as e:
+            ctx.__exit__(type(e), e, None)
+            raise
 
     def test_outbox_rescheduling(self) -> None:
         with patch(
@@ -710,393 +558,6 @@ class RegionOutboxTest(TransactionTestCase):
             # We expect the shard to be drained if at *least* one scheduled
             # message is in the past.
             assert RegionOutbox.objects.count() == 0
-
-    def test_process_coalesced_explicit_ordering(self) -> None:
-        """Test that coalesced messages are processed in the correct order"""
-        outboxes = []
-        for i in range(5):
-            outbox = OrganizationMember(id=i + 1, organization_id=3, user_id=1).outbox_for_update()
-            outboxes.append(outbox)
-
-        with outbox_context(flush=False):
-            for outbox in reversed(outboxes):  # Save in reverse order
-                outbox.save()
-
-        processed_ids = []
-        original_send_signal = RegionOutbox.send_signal
-
-        def mock_send_signal(self: Any) -> None:
-            processed_ids.append(self.id)
-            original_send_signal(self)
-
-        with patch.object(RegionOutbox, "send_signal", mock_send_signal):
-            # Process all messages in the shard
-            outboxes[0].drain_shard(flush_all=True)
-
-        # Verify messages were processed in order by ID
-        assert processed_ids == sorted(processed_ids)
-        assert RegionOutbox.objects.count() == 0
-
-    def test_process_coalesced_lock_acquisition(self) -> None:
-        """Test lock acquisition behavior during coalesced processing"""
-        outbox1 = OrganizationMember(id=1, organization_id=3, user_id=1).outbox_for_update()
-        outbox2 = OrganizationMember(
-            id=2, organization_id=4, user_id=2
-        ).outbox_for_update()  # Different organization and user
-        outbox1.shard_identifier = 100
-        outbox2.shard_identifier = 200  # Different shard_identifier
-
-        with outbox_context(flush=False):
-            outbox1.save()
-            outbox2.save()
-
-        # Create barriers to synchronize threads
-        barrier = threading.Barrier(2, timeout=5)
-        lock_acquired = threading.Event()
-        process_completed = threading.Event()
-
-        def process_thread_1() -> None:
-            try:
-                with outbox1.process_coalesced(is_synchronous_flush=True) as coalesced:
-                    if coalesced:
-                        lock_acquired.set()
-                        barrier.wait()  # Signal thread 2 to try processing
-                        process_completed.wait()  # Wait for thread 2 to attempt processing
-                        coalesced.send_signal()
-            finally:
-                for conn in connections.all():
-                    conn.close()
-
-        def process_thread_2() -> None:
-            try:
-                barrier.wait()  # Wait for thread 1 to acquire lock
-                assert lock_acquired.is_set()  # Ensure thread 1 has the lock
-                with outbox2.process_coalesced(is_synchronous_flush=True) as coalesced:
-                    if coalesced:
-                        coalesced.send_signal()
-                process_completed.set()
-            finally:
-                for conn in connections.all():
-                    conn.close()
-
-        t1 = threading.Thread(target=process_thread_1)
-        t2 = threading.Thread(target=process_thread_2)
-
-        t1.start()
-        t2.start()
-        t1.join()
-        t2.join()
-
-        # Verify final state
-        assert RegionOutbox.objects.count() == 0
-
-    def test_process_coalesced_stable_snapshot(self) -> None:
-        """Test that deletion uses a stable snapshot of coalesced messages"""
-        outbox1 = OrganizationMember(id=1, organization_id=3, user_id=1).outbox_for_update()
-        # Explicitly set all coalescing key fields for outbox1
-        outbox1.shard_scope = OutboxScope.ORGANIZATION_SCOPE
-        outbox1.shard_identifier = 100
-        outbox1.category = OutboxCategory.ORGANIZATION_MEMBER_UPDATE
-        outbox1.object_identifier = 1
-        outbox1.scheduled_for = django_timezone.now() - timedelta(minutes=1)  # Past
-
-        with outbox_context(flush=False):
-            outbox1.save()
-
-        barrier = threading.Barrier(2, timeout=5)
-        snapshot_taken = threading.Event()
-        processing_started = threading.Event()
-
-        def concurrent_insert() -> None:
-            try:
-                barrier.wait()  # Wait for main thread to start processing
-                snapshot_taken.wait()  # Wait for snapshot to be taken
-                # Add new message after snapshot with different organization and member IDs
-                with outbox_context(flush=False):
-                    outbox2 = OrganizationMember(
-                        id=2, organization_id=4, user_id=2
-                    ).outbox_for_update()
-                    # Explicitly set all coalescing key fields for outbox2 to be different from outbox1
-                    outbox2.shard_scope = OutboxScope.ORGANIZATION_SCOPE  # Same scope as outbox1
-                    outbox2.shard_identifier = 200  # Different organization ID
-                    outbox2.category = OutboxCategory.ORGANIZATION_MEMBER_UPDATE  # Same category
-                    outbox2.object_identifier = 2  # Different member ID
-                    outbox2.scheduled_for = django_timezone.now() + timedelta(hours=1)
-                    outbox2.save()
-                barrier.wait()  # Signal main thread to continue
-            finally:
-                for conn in connections.all():
-                    conn.close()
-
-        # Patch drain_shard to prevent background processing during test
-        with patch("sentry.hybridcloud.models.outbox.OutboxBase.drain_shard"):
-            t = threading.Thread(target=concurrent_insert)
-            t.start()
-
-            # Process outbox1 first
-            using = router.db_for_write(RegionOutbox)
-            with connections[using].cursor() as cursor:
-                cursor.execute("SET TRANSACTION ISOLATION LEVEL SERIALIZABLE")
-                cursor.close()
-
-            with outbox1.process_coalesced(is_synchronous_flush=True) as coalesced:
-                if not processing_started.is_set():
-                    processing_started.set()
-                    snapshot_taken.set()  # Signal that we've taken the snapshot
-                    barrier.wait()  # Wait for concurrent insert
-                    barrier.wait()  # Wait for concurrent insert to complete
-                if coalesced:
-                    coalesced.send_signal()
-
-            t.join()
-
-            # Verify that outbox1 was deleted but outbox2 remains
-            assert not RegionOutbox.objects.filter(
-                id=outbox1.id
-            ).exists()  # outbox1 should be deleted
-            # Verify outbox2 by its unique coalescing keys instead of hardcoded ID
-            assert RegionOutbox.objects.filter(
-                shard_scope=OutboxScope.ORGANIZATION_SCOPE,
-                shard_identifier=200,
-                category=OutboxCategory.ORGANIZATION_MEMBER_UPDATE,
-                object_identifier=2,
-            ).exists()  # outbox2 should remain
-
-    def test_process_coalesced_transaction_boundaries(self) -> None:
-        """Test transaction boundaries during coalesced processing"""
-        outbox1 = OrganizationMember(id=1, organization_id=3, user_id=1).outbox_for_update()
-
-        with outbox_context(flush=False):
-            outbox1.save()
-
-        transaction_state = []
-        original_send_signal = RegionOutbox.send_signal
-
-        def mock_send_signal(self: Any) -> None:
-            # Check if we're in a transaction during signal sending
-            using = router.db_for_write(type(self))
-            transaction_state.append(transaction.get_connection(using=using).in_atomic_block)
-            original_send_signal(self)
-
-        with patch.object(RegionOutbox, "send_signal", mock_send_signal):
-            with outbox1.process_coalesced(is_synchronous_flush=True) as coalesced:
-                if coalesced:
-                    coalesced.send_signal()
-
-        # Verify signal was sent outside transaction
-        assert not any(transaction_state), "Signal was sent within a transaction"
-        assert RegionOutbox.objects.count() == 0
-
-    def test_process_coalesced_scheduled_for(self) -> None:
-        """
-        Test that process_coalesced() updates scheduled_for for all messages in a group.
-        """
-
-        # Use a fixed point in time for determinism.
-        fixed_time = datetime(2022, 1, 1, 0, 0, 0, tzinfo=timezone.utc)
-        with freeze_time(fixed_time):
-            # Create two outbox messages sharing the same coalescing keys.
-            # (For example, using OrganizationMember.outbox_for_update() if that is how your app creates them.)
-            outbox1 = OrganizationMember(id=1, organization_id=3, user_id=1).outbox_for_update()
-            outbox2 = OrganizationMember(id=2, organization_id=3, user_id=1).outbox_for_update()
-            # Explicitly set all coalescing key fields so that they are grouped together.
-            for outbox in (outbox1, outbox2):
-                outbox.shard_scope = OutboxScope.ORGANIZATION_SCOPE
-                outbox.shard_identifier = 100
-                outbox.category = OutboxCategory.ORGANIZATION_MEMBER_UPDATE
-                outbox.object_identifier = 1
-                # Ensure at least one message is scheduled in the past so that they are eligible.
-                outbox.scheduled_for = fixed_time - timedelta(minutes=1)
-
-            with outbox_context(flush=False):
-                outbox1.save()
-                outbox2.save()
-
-            # Call process_coalesced() to force the reservation step.
-            with outbox1.process_coalesced(is_synchronous_flush=True):
-                # At this point the reservation (UPDATE of scheduled_for) has been committed.
-                # Retrieve the group of messages using their coalescing key fields.
-                group = list(
-                    RegionOutbox.objects.filter(
-                        shard_scope=outbox1.shard_scope,
-                        shard_identifier=outbox1.shard_identifier,
-                        category=outbox1.category,
-                        object_identifier=outbox1.object_identifier,
-                    ).order_by("id")
-                )
-
-                # We expect both messages (or however many are in the group) to have their scheduled_for updated.
-                # The new scheduled_for should be roughly fixed_time + 1 hour.
-                expected = fixed_time + timedelta(hours=1)
-
-                # Allow a small tolerance (a few seconds) in timing.
-                tolerance = 5  # seconds
-                for msg in group:
-                    diff = abs((msg.scheduled_for - expected).total_seconds())
-                    assert (
-                        diff < tolerance
-                    ), f"Message {msg.id} had scheduled_for {msg.scheduled_for}, expected ~{expected}"
-
-    def test_process_coalesced_lock_error_handling(self) -> None:
-        """Test error handling during lock acquisition"""
-        outbox1 = OrganizationMember(id=1, organization_id=3, user_id=1).outbox_for_update()
-
-        with outbox_context(flush=False):
-            outbox1.save()
-
-        def mock_select_for_update(*args: Any, **kwargs: Any) -> Any:
-            raise OperationalError("could not obtain lock")
-
-        # Mock the select_for_update at the queryset level
-        with patch(
-            "django.db.models.query.QuerySet.select_for_update", side_effect=mock_select_for_update
-        ):
-            # Should handle lock error gracefully
-            with outbox1.process_coalesced(is_synchronous_flush=True) as coalesced:
-                assert coalesced is None
-
-        # Original message should still exist
-        assert RegionOutbox.objects.filter(id=outbox1.id).exists()
-
-    def test_high_concurrency_reservation(self) -> None:
-        """
-        Simulate high-concurrency processing of a coalesced group.
-
-        Create a group of outbox messages that share the same coalescing keys.
-        Spawn several threads that call process_coalesced() concurrently.
-        Verify that exactly one thread obtains the reservation (non-None result)
-        while all others yield None.
-        """
-        from threading import Barrier, Thread
-
-        # Create a coalesced group – two messages with identical coalescing fields.
-        outbox1 = OrganizationMember(id=1, organization_id=3, user_id=1).outbox_for_update()
-        outbox2 = OrganizationMember(id=2, organization_id=3, user_id=1).outbox_for_update()
-        for outbox in (outbox1, outbox2):
-            outbox.shard_scope = OutboxScope.ORGANIZATION_SCOPE
-            outbox.shard_identifier = 100
-            outbox.category = OutboxCategory.ORGANIZATION_MEMBER_UPDATE
-            outbox.object_identifier = 1
-            # Make sure they are eligible (set scheduled_for in the past)
-            outbox.scheduled_for = django_timezone.now() - timedelta(minutes=1)
-
-        with outbox_context(flush=False):
-            outbox1.save()
-            outbox2.save()
-
-        THREAD_COUNT = 10
-        barrier = Barrier(THREAD_COUNT)
-        results = []  # we'll store each thread's result here
-
-        def worker() -> None:
-            # wait for all threads to be ready
-            barrier.wait()
-            # Each thread will attempt to process the same group.
-            with outbox1.process_coalesced(is_synchronous_flush=True) as result:
-                results.append(result)
-
-        threads = [Thread(target=worker) for _ in range(THREAD_COUNT)]
-        for t in threads:
-            t.start()
-        for t in threads:
-            t.join()
-
-        # Count how many threads got a valid (non-None) reservation.
-        non_none = [res for res in results if res is not None]
-        assert len(non_none) == 1, (
-            f"Expected exactly one thread with a reservation, got {len(non_none)}. "
-            f"Results: {results}"
-        )
-
-    def test_scheduled_for_prevents_reselection(self) -> None:
-        """
-        Verify that scheduled_for is updated to a future timestamp such that the reserved
-        messages are no longer selected for processing by subsequent queries.
-
-        The test creates a coalesced group, then calls process_coalesced() to update scheduled_for.
-        Afterwards it queries for messages that would be eligible for processing (with scheduled_for <= now)
-        and asserts that the group is not returned.
-        """
-        fixed_time = datetime(2022, 1, 1, 0, 0, 0, tzinfo=timezone.utc)
-        with freeze_time(fixed_time):
-            # Create two messages with the same coalescing keys.
-            outbox1 = OrganizationMember(id=1, organization_id=3, user_id=1).outbox_for_update()
-            outbox2 = OrganizationMember(id=2, organization_id=3, user_id=1).outbox_for_update()
-            for outbox in (outbox1, outbox2):
-                outbox.shard_scope = OutboxScope.ORGANIZATION_SCOPE
-                outbox.shard_identifier = 100
-                outbox.category = OutboxCategory.ORGANIZATION_MEMBER_UPDATE
-                outbox.object_identifier = 1
-                # Set scheduled_for in the past so that they are eligible.
-                outbox.scheduled_for = fixed_time - timedelta(minutes=1)
-
-            with outbox_context(flush=False):
-                outbox1.save()
-                outbox2.save()
-
-        # Execute process_coalesced; when it executes the reservation update,
-        # scheduled_for will be set to fixed_time + 1 hour.
-        with outbox1.process_coalesced(is_synchronous_flush=True):
-            # Reservation has been committed at this point.
-            # Now, try to re-select messages from the same group that are eligible for processing.
-            reserved = list(
-                RegionOutbox.objects.filter(
-                    shard_scope=outbox1.shard_scope,
-                    shard_identifier=outbox1.shard_identifier,
-                    category=outbox1.category,
-                    object_identifier=outbox1.object_identifier,
-                    scheduled_for__lte=django_timezone.now(),
-                )
-            )
-
-        # Since the scheduled_for was updated to fixed_time + 1 hour, no message should be eligible.
-        assert (
-            len(reserved) == 0
-        ), "Expected no re-selection of reserved messages, but found some eligible for processing."
-
-    def test_process_coalesced_cleanup_on_signal_failure_no_deletion(self) -> None:
-        """
-        Test that if coalesced.send_signal() raises an exception, the cleanup
-        does not delete the reserved (coalesced) message.
-
-        In our current behavior the reserved message remains in the database.
-        This test confirms that after a failure, the message is still present.
-        """
-        fixed_time = datetime(2022, 1, 1, 0, 0, 0, tzinfo=timezone.utc)
-        with freeze_time(fixed_time):
-            # Create a message that forms a coalesced group (with a single message).
-            outbox = OrganizationMember(id=1, organization_id=3, user_id=1).outbox_for_update()
-            outbox.shard_scope = OutboxScope.ORGANIZATION_SCOPE
-            outbox.shard_identifier = 100
-            outbox.category = OutboxCategory.ORGANIZATION_MEMBER_UPDATE
-            outbox.object_identifier = 1
-            # Set scheduled_for in the past to mark them eligible.
-            outbox.scheduled_for = django_timezone.now() - timedelta(minutes=1)
-
-            with outbox_context(flush=False):
-                outbox.save()
-
-        # Patch the send_signal method so that it raises an exception.
-        # The expected behavior currently is that the cleanup phase does not delete the message.
-        from unittest.mock import patch
-
-        with patch.object(RegionOutbox, "send_signal", side_effect=ValueError("TEST ERROR")):
-            with pytest.raises(ValueError, match="TEST ERROR"):
-                with outbox.process_coalesced(is_synchronous_flush=True) as coalesced:
-                    if coalesced:
-                        coalesced.send_signal()
-
-        # After the exception, check that the message has NOT been deleted.
-        remaining = RegionOutbox.objects.filter(
-            shard_scope=outbox.shard_scope,
-            shard_identifier=outbox.shard_identifier,
-            category=outbox.category,
-            object_identifier=outbox.object_identifier,
-        ).first()
-        # In the current behavior, the reserved message remains in the database.
-        assert remaining is not None, "Expected the reserved message to remain on error"
-
-        assert remaining.scheduled_for == outbox.scheduled_for
 
 
 class TestOutboxesManager(TestCase):
